requirements.txt
flask
flask-sqlalchemy
flask-migrate
python-dotenv
psycopg
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openai
